remove print in grpc_client and update more test functions in grpc_clien_test
