+++ quarkchain/cluster/grpc_client.py
@@ -2,10 +2,8 @@ import logging
 import sys
 import grpc
 
-sys.path.append("quarkchain/cluster")
-
-import grpc_client_pb2_grpc
-import grpc_client_pb2
+from quarkchain.cluster import grpc_client_pb2_grpc
+from quarkchain.cluster import grpc_client_pb2
 
 
 class GrpcClient(object):
@@ -26,15 +24,12 @@ class GrpcClient(object):
         try:
             response = self.stub.SetRootChainConfirmedBlock(request)
         except Exception as e:
-            print(str(e))
+            # print(str(e))
             return False
 
-        print("-------------- SetRootChainConfirmedBlock --------------")
-        print(response.status.code, response.status.message)
         if response.status.code == 0:
             return True
         else:
-            print("response code status abnormal.", response.status.code)
             return False
 
 
@@ -46,7 +41,7 @@ def run(channel):
 def main():
     logging.basicConfig()
     if len(sys.argv) < 3:
-        print("Usage--python client.py host port")
+        print("Usage--python grpc_client.py host port")
     else:
         host = sys.argv[1]
         port = sys.argv[2]
